+++ lib/screen/category_screen/imports.dart
@@ -1,0 +1,5 @@
+export 'package:flutter/material.dart';
+export '../../widgets/appBar.dart';
+export 'widgets/categories_recommendation.dart';
+export '../../../widgets/recommendation.dart';
+export '../../../widgets/title_heading.dart';
